hiver_middleware

Module middleware

Middleware module 中间件模块

Equivalent to Spring Boot / 等价于 Spring Boot

Filter, OncePerRequestFilter

HandlerInterceptor

WebMvcConfigurer

Overview / 概述

This module re-exports the Middleware trait and Next type from hiver-router. All middleware implementations should use these types for compatibility.

此模块从hiver-router重新导出Middleware trait和Next类型。 所有中间件实现应使用这些类型以确保兼容性。

Structs

MiddlewareStack

Middleware stack 中间件栈

Next

Next middleware in the chain 链中的下一个中间件

Request

HTTP Request HTTP 请求

Response

HTTP Response / HTTP 响应

Enums

Error

HTTP Error type HTTP 错误类型

Traits

Middleware

Middleware trait 中间件trait

Type Aliases

Result

Result type for HTTP operations HTTP操作的Result类型